Moved ParallelFileMerger to its own file.
